Student-Projects/group17/requirements.txt
python-telegram-bot==21.10
groq==0.16.0
python-dotenv==1.0.1
pypdf==4.0.1
python-docx==1.1.0
ebooklib==0.18
beautifulsoup4==4.12.3
requests==2.31.0
aiosqlite==0.19.0

